+++ build.sbt
@@ -6,7 +6,8 @@ val tools = Build.tools
 val toolsJS = Build.toolsJS
 val jsEnvs = Build.jsEnvs
 val jsEnvsTestKit = Build.jsEnvsTestKit
-val jsEnvsTestSuite = Build.jsEnvsTestSuite
+val nodeJSEnv = Build.nodeJSEnv
+val jsdomNodeJSEnv = Build.jsdomNodeJSEnv
 val testAdapter = Build.testAdapter
 val sbtPlugin = Build.plugin
 val javalanglib = Build.javalanglib
