+++ nodejs-env/src/test/scala/org/scalajs/jsenv/nodejs/NodeJSTest.scala
@@ -1,6 +1,6 @@
-package org.scalajs.jsenv.test
+package org.scalajs.jsenv.nodejs
 
-import org.scalajs.jsenv.nodejs.NodeJSEnv
+import org.scalajs.jsenv.test._
 
 import org.junit.Test
 import org.junit.Assert._
