+++ project/build.sbt
@@ -27,6 +27,8 @@ unmanagedSourceDirectories in Compile ++= {
     root / "tools/shared/src/main/scala",
     root / "tools/jvm/src/main/scala",
     root / "js-envs/src/main/scala",
+    root / "nodejs-env/src/main/scala",
+    root / "jsdom-nodejs-env/src/main/scala",
     root / "test-adapter/src/main/scala",
     root / "sbt-plugin/src/main/scala",
     root / "jsdependencies-core/src/main/scala",
