+++ sbt-plugin-test/project/build.sbt
@@ -1,2 +1,5 @@
 addSbtPlugin("org.scala-js" % "sbt-scalajs" %
   org.scalajs.core.ir.ScalaJSVersions.current)
+
+libraryDependencies +=
+  "org.scala-js" %% "scalajs-jsdom-nodejs-env" % org.scalajs.core.ir.ScalaJSVersions.current
